Dirty integration of UniqueIDPool.
  In case of all entities are requiring an unique id passed into
their constructors i created UniqueIDPool. But the objective to
integrate it  in the project's architecure without even more mess up
is very complex, so i just inherit IDataStorage interface
directly from UniqueIDPool. So IDataStorage is becoming to
look like a "superclass" more and more, which is a sign of poor
architecture planning on the start.

  Also in this commit i modified a test func integrated
directly into GameEngine and the "shooting" function
_spawn_a_projectile_correspond_to_the_shoot_action(...)
from the BotAICore to fit new requirments